Support for known proteins
The protheaders parameter may have a prefix indicating if any or all proteins in the matching list should be considered. If the prefix is 'known' then the it is returned if any of the proteins match the given headers whereas for other prefixes all proteins in the list should match the given headers. Input example for protheaders: known:^ENSP;?^decoy_ENSP or var:^lncRNA;^decoy_lncRNA.